Moved constant stuff to the CONST object, and added typeof declarations on the PIXI object.
The thing is, the "CONST" object doesn't exist on the PIXI object, all its properties are added directly to the PIXI object.
But i can understand that for readability, it is nicer to have everything documented under the CONST module.
So what I've done is to document everything under the CONST object, and then add typeof declarations, pointing to CONST on the PIXI object.
This avoids duplicate declarations (because typeof is essentially a pointer), and everything is accessible on the PIXI object, to reflect how the PIXI object actually looks after initialization.